+++ changelog.md
@@ -29,3 +29,6 @@
 ### 2020-12-21
 - fixed issues with trainer module caused by incorrect loss function, made trainer as extension of module class
 - fixed checkpoint issues, earlier was recording losses instead of optimizers hence fixed it.
+
+### 2020-12-22
+- feat added dcgan trainer for training DCGAN